Implemented buttons and time increase
now has back buttons and the time system. clicking on the door opens the time increase ui and it generates new emails when you progress time. Multiple changes have been made to the sequencing of the emails to accomdate this.

Currently the submit button remains even after a build has been submitted as it is still in the ongoing builds array. I have yet to come up with a solution for this